+++ src/Ledger/RolloverLedger/hooks/index.js
@@ -1,4 +1,5 @@
 export * from './useRolloverData';
+export * from './useRolloverErrorHandler';
 export * from './useRolloverFiscalYear';
 export * from './useUnpaidInvoices';
 export * from './useUnpaidInvoicesExport';